# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Grove Street - Arabic Gaming Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ads:
    get:
      operationId: listAds
      tags: [ads]
      summary: List active ads
      responses:
        "200":
          description: Ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"
    post:
      operationId: createAd
      tags: [ads]
      summary: Create ad (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
  /ads/{id}:
    patch:
      operationId: updateAd
      tags: [ads]
      summary: Update ad (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"
    delete:
      operationId: deleteAd
      tags: [ads]
      summary: Delete ad (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
...
  /admin/settings:
    get:
      operationId: getSettings
      tags: [admin]
      summary: Get site settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
    patch:
      operationId: updateSettings
      tags: [admin]
      summary: Update site settings (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SiteSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SiteSettings"
components:
  schemas:
...
    Ad:
      type: object
      required: [id, title, isActive, order]
      properties:
        id:            { type: integer }
        title:         { type: string }
        titleAr:       { type: string }
        description:   { type: string }
        descriptionAr: { type: string }
        imageUrl:      { type: string }
        linkUrl:       { type: string }
        buttonText:    { type: string }
        buttonTextAr:  { type: string }
        isActive:      { type: boolean }
        order:         { type: integer }
    AdInput:
      type: object
      required: [title]
      properties:
        title:         { type: string }
        titleAr:       { type: string }
        description:   { type: string }
        descriptionAr: { type: string }
        imageUrl:      { type: string }
        linkUrl:       { type: string }
        buttonText:    { type: string }
        buttonTextAr:  { type: string }
        isActive:      { type: boolean }
        order:         { type: integer }
    AdUpdate:
      type: object
      properties:
        title:         { type: string }
        titleAr:       { type: string }
        description:   { type: string }
        descriptionAr: { type: string }
        imageUrl:      { type: string }
        linkUrl:       { type: string }
        buttonText:    { type: string }
        buttonTextAr:  { type: string }
        isActive:      { type: boolean }
        order:         { type: integer }
...
    SiteSettings:
      type: object
      required: [siteName, maintenanceMode]
      properties:
        siteName:        { type: string }
        siteDescription: { type: string }
        maintenanceMode: { type: boolean }
        contactEmail:    { type: string }
        contactPhone:    { type: string }
        socialLinks:     { type: string }
    SiteSettingsUpdate:
      type: object
      properties:
        siteName:        { type: string }
        siteDescription: { type: string }
        maintenanceMode: { type: boolean }
        contactEmail:    { type: string }
        contactPhone:    { type: string }
        socialLinks:     { type: string }